+++ lib/graphql/schema.rb
@@ -221,7 +221,7 @@ module GraphQL
       end
 
       def default_filter
-        GraphQL::Filter.new(except: default_mask)
+        GraphQL::Filter.new(except: default_mask, silence_deprecation_warning: true)
       end
 
       def default_mask(new_mask = nil)
@@ -799,11 +799,7 @@ module GraphQL
             end
 
             if resolved_type.nil? || (resolved_type.is_a?(Module) && resolved_type.respond_to?(:kind))
-              if resolved_value
-                [resolved_type, resolved_value]
-              else
-                resolved_type
-              end
+              [resolved_type, resolved_value]
             else
               raise ".resolve_type should return a type definition, but got #{resolved_type.inspect} (#{resolved_type.class}) from `resolve_type(#{type}, #{obj}, #{ctx})`"
             end